refactor: Move card parts to card module
